everything works and looks fine, but we want to have one hook in RegisterAccount.js for all
