+++ client/static/index.html
@@ -149,7 +149,7 @@
     }
     label {
       padding: 9px 0 9px 100px;
-      width: 280px;
+      width: 290px;
       text-align: right;
     }
     .profileEdit {
@@ -178,6 +178,20 @@
 
 input:focus:invalid {
   outline: none;
+}
+
+input {
+  border-radius: 5px;
+  margin: 7px;
+
+}
+
+button {
+  margin-top: 30px;
+    background: #ffe6cc;
+    border-radius: 10px;
+    width: 200px;
+    height: 80%;
 }
 
 /* This is the style of our error messages */
